Refactor Datamodel, clean up UserController and add PanelUserController

+++ api/Ableport.API.REST/Controllers/PanelUser.cs
@@ -1,16 +1,17 @@
 using Ableport.API.REST.DataModel;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Namotion.Reflection;
 
 namespace Ableport.API.REST.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class UserController : ControllerBase
+    public class PanelUserController : ControllerBase
     {
         private readonly AbleportContext _context;
 
-        public UserController(AbleportContext context)
+        public PanelUserController(AbleportContext context)
         {
             _context = context;
         }
@@ -33,6 +34,18 @@ namespace Ableport.API.REST.Controllers
         [HttpPost]
         public async Task<ActionResult<AbleportUser>> CreateUser(AbleportUser user)
         {
+            // Insert default data if not included
+            if (user.PanelUserData == null)
+            {
+                var defaultUserData = new PanelUserData()
+                {
+                    UserId = user.Id,
+                    AllowCommercialPanels = false,
+                    ContactPreference = "No contact",
+                };
+                user.PanelUserData = defaultUserData;
+            }
+            
             _context.Users.Add(user);
             await _context.SaveChangesAsync();
 
@@ -41,11 +54,11 @@ namespace Ableport.API.REST.Controllers
 
         // PUT: api/user/{id}
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateUser(string id, AbleportUser user)
+        public async Task<IActionResult> UpdateUserData(string id, PanelUserData data)
         {
-            if (id != user.Id) return BadRequest("User IDs do not match");
+            if (id != data.UserId) return BadRequest("User IDs do not match");
 
-            _context.Entry(user).State = EntityState.Modified;
+            _context.Entry(data).State = EntityState.Modified;
 
             try
             {
@@ -68,9 +81,12 @@ namespace Ableport.API.REST.Controllers
         public async Task<IActionResult> DeleteUser(string id)
         {
             var user = await _context.Users.FindAsync(id);
-
+            var hasPanelData = _context.Users.Where(d => d.Id == id)
+                .Any(u => u.PanelUserData != null);
+            
             if (user == null) return NotFound();
-
+            if (!hasPanelData) return BadRequest("Not a PanelUser");
+            
             _context.Users.Remove(user);
             await _context.SaveChangesAsync();
